tests/wpt/web-platform-tests/sanitizer-api/sanitizer-config.tentative.html

<!DOCTYPE html>
<html>
<head>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script src="support/testcases.sub.js"></script>
</head>

<body>
<script>
    test(t => {
      let s = new Sanitizer();
      assert_true(s instanceof Sanitizer);
      assert_object_equals(s.creationOptions, {});
    }, "SanitizerAPI creator without config.");

    test(t => {
      let s = new Sanitizer({});
      assert_true(s instanceof Sanitizer);
      assert_object_equals(s.creationOptions, {});
    }, "SanitizerAPI creator with empty config.");

    test(t => {
      let s = new Sanitizer(null);
      assert_true(s instanceof Sanitizer);
      assert_object_equals(s.creationOptions, {});
    }, "SanitizerAPI creator with null as config.");

    test(t => {
      let s = new Sanitizer(undefined);
      assert_true(s instanceof Sanitizer);
      assert_object_equals(s.creationOptions, {});
    }, "SanitizerAPI creator with undefined as config.");

    test(t => {
      let s = new Sanitizer({testConfig: [1,2,3], attr: ["test", "i", "am"]});
      assert_true(s instanceof Sanitizer);
      assert_object_equals(s.creationOptions, {});
    }, "SanitizerAPI creator with config ignore unknown values.");

    test(t => {
      let options = {};
      let s = new Sanitizer(options);
      assert_true(s instanceof Sanitizer);
      assert_object_equals(s.creationOptions, {});

      options.testConfig = [1,2,3];
      assert_object_equals(s.creationOptions, {});

      s.creationOptions = {testConfig: [1,2,3]};
      assert_object_equals(s.creationOptions, {});

      s.creationOptions['testConfig'] = [1,2,3];
      assert_object_equals(s.creationOptions, {});
    }, "SanitizerAPI config is not editable.");

</script>
</body>
</html>
